get-ssl-ali/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "aliyun-ssl-manager"
version = "0.1.0"
description = "Automated SSL certificate management via ACME (Let's Encrypt) with Alibaba Cloud DNS-01 validation and SSH deployment"
readme = "../README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "MQ", email = "243329721@qq.com"},
]
keywords = ["ssl", "certificate", "acme", "letsencrypt", "aliyun", "alibaba-cloud", "dns-01", "nginx", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]
dependencies = [
    "alibabacloud-cas20200407>=3.0.0",
    "alibabacloud-alidns20150109>=3.0.0",
    "alibabacloud-tea-openapi>=0.3.0",
    "paramiko>=3.0.0",
    "pyyaml>=6.0",
    "acme>=3.0.0",
    "josepy>=2.0.0",
    "cryptography>=42.0.0,<45.0.0",
]

[project.urls]
Homepage = "https://github.com/qq243329721/get-ssl-ali"
Repository = "https://github.com/qq243329721/get-ssl-ali"
Issues = "https://github.com/qq243329721/get-ssl-ali/issues"

[project.scripts]
aliyun-ssl = "aliyun_ssl_manager.cli:main"

[tool.setuptools.packages.find]
where = ["scripts"]
